requirements.txt
# Ingestion API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
confluent-kafka==2.3.0
requests==2.31.0

# Processing Engine
redis==5.0.1
pyspark==3.5.0
delta-spark==3.1.0

# Dashboard & Database
streamlit==1.31.0
pandas==2.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25